consumer/requirements.txt
pika==1.2.1
Pillow==8.2.0
requests==2.25.1